+++ IotEdgeGateway/src/gateway/gateway_core.cpp
@@ -18,7 +18,7 @@
 #include "core/common/logger/logger.hpp"
 #include "core/common/utils/json_utils.hpp"
 #include "core/common/utils/time_utils.hpp"
-#include "core/control/rule_engine/rule_engine.hpp"
+#include "core/control/rule_engine.hpp"
 #include "core/device/manager/device_manager.hpp"
 #include "core/device/protocol_adapters/mqtt_adapter/mqtt_adapter.hpp"
 #include "services/system_services/update/update_manager.hpp"